Rename: 폴더 이름 변경 Modal to Dialog
