+++ src/components/Dialog/DiaryDeleteDialog.module.scss
